Refactor & adjustments of StabilityAIAPIInpaintingModelBlockV1
