+++ inference/core/workflows/core_steps/loader.py
@@ -57,8 +57,8 @@ from inference.core.workflows.core_steps.models.foundation.openai.v1 import (
 from inference.core.workflows.core_steps.models.foundation.segment_anything2.v1 import (
     SegmentAnything2BlockV1,
 )
-from inference.core.workflows.core_steps.models.foundation.flux1_inpaint.v1 import (
-    Flux1InpaintingBlockV1,
+from inference.core.workflows.core_steps.models.foundation.flux1_inpaint import (
+    StabilityAIAPIInpaintingModelBlockV1,
 )
 from inference.core.workflows.core_steps.models.foundation.yolo_world.v1 import (
     YoloWorldModelBlockV1,
@@ -269,7 +269,7 @@ def load_blocks() -> List[Type[WorkflowBlock]]:
         SIFTComparisonBlockV1,
         SIFTBlockV1,
         TemplateMatchingBlockV1,
-        Flux1InpaintingBlockV1,
+        StabilityAIAPIInpaintingModelBlockV1,
     ]
